+++ aclocal.m4
@@ -149,82 +149,6 @@ AC_DEFUN([AX_CXX_COMPILE_STDCXX_11], [dnl
   fi
 ])
 
-dnl ----------------------------------------------------------------------------
-dnl check for the required PETSc library
-dnl ----------------------------------------------------------------------------
-AC_DEFUN([ACX_PETSc], [
-AC_REQUIRE([ACX_BLAS])
-BLAS_LIBS="$BLAS_LIBS $FLIBS"
-AC_REQUIRE([ACX_LAPACK])
-LAPACK_LIBS="$LAPACK_LIBS $BLAS_LIBS"
-AC_PATH_XTRA
-
-if test "x$PETSC_DIR" == "x"; then
-  AC_MSG_WARN( [No PETSC_DIR set - trying to autodetect] )
-
-  # Try to identify the obvious choice
-  if test -f /usr/lib/petsc/include/petscversion.h ; then
-    export PETSC_DIR=/usr/lib/petsc
-  elif test -f /usr/include/petscversion.h ; then
-    export PETSC_DIR=/usr
-  elif test -f /usr/local/include/petscversion.h ; then
-    export PETSC_DIR=/usr/local
-  fi
-fi
-# Check again incase we failed.
-if test "x$PETSC_DIR" == "x"; then 
-  AC_MSG_WARN( [No PETSC_DIR set - do you need to load a petsc module?] )
-  AC_MSG_ERROR( [You need to set PETSC_DIR to point at your PETSc installation... exiting] )
-fi
-AC_MSG_NOTICE([Using PETSC_DIR=$PETSC_DIR])
-
-PETSC_LINK_LIBS=`make -s -f petsc_makefile getlinklibs`
-LIBS="$PETSC_LINK_LIBS $LIBS"
-
-PETSC_INCLUDE_FLAGS=`make -s -f petsc_makefile getincludedirs`
-CPPFLAGS="$CPPFLAGS $PETSC_INCLUDE_FLAGS"
-FCFLAGS="$FCFLAGS $PETSC_INCLUDE_FLAGS"
-
-# Horrible hacks needed for cx1
-# Somehow /apps/intel/ict/mpi/3.1.038/lib64 gets given as /apps/intel/ict/mpi/3.1.038/lib/64
-# maybe the directory got moved after building petsc? Anyhow next time we
-# request a new petsc package on cx1, this hack can be removed - and we
-# can check if the new packages passes the test without any hacks.
-fixedLIBS=`echo $LIBS |sed 's@/apps/intel/ict/mpi/3.1.038/lib/64@/apps/intel/ict/mpi/3.1.038/lib64@g'`
-
-if test ! "$LIBS" == "$fixedLIBS"; then
-  # more fixes needed:
-  # also -lmpichcxx got mangled to -lmpichxx
-  LIBS=`echo $fixedLIBS |sed 's@mpichxx@mpichcxx@g'`
-  # remove -lPEPCF90, -lpromfei and -lprometheus
-  LIBS=`echo $LIBS |sed 's@-lPEPCF90@@g'`
-  LIBS=`echo $LIBS |sed 's@-lpromfei@@g'`
-  LIBS=`echo $LIBS |sed 's@-lprometheus@@g'`
-fi
-
-# finally check we have the right petsc version
-AC_COMPUTE_INT(PETSC_VERSION_MAJOR, "PETSC_VERSION_MAJOR", [#include "petscversion.h"], 
-  [AC_MSG_ERROR([Unknown petsc major version])])
-AC_COMPUTE_INT(PETSC_VERSION_MINOR, "PETSC_VERSION_MINOR", [#include "petscversion.h"], 
-  [AC_MSG_ERROR([Unknown petsc minor version])])
-AC_MSG_NOTICE([Detected PETSc version "$PETSC_VERSION_MAJOR"."$PETSC_VERSION_MINOR"])
-# if major<3 or minor<1
-if test "0$PETSC_VERSION_MAJOR" -lt 3 -o "0$PETSC_VERSION_MINOR" -lt 1; then
-  AC_MSG_ERROR([Fluidity needs PETSc version >=3.1])
-fi
-
-AC_DEFINE(HAVE_PETSC,1,[Define if you have the PETSc library.])
-
-# define HAVE_PETSC33 for use in the Makefiles (including petsc's makefiles
-# would require having PETSC_DIR+PETSC_ARCH set correctly for every make)
-if test "0$PETSC_VERSION_MINOR" -ge 3; then
-  HAVE_PETSC33=yes
-else
-  HAVE_PETSC33=no
-fi
-AC_SUBST(HAVE_PETSC33)
-
-])dnl ACX_PETSc
 dnl @synopsis ACX_BLAS([ACTION-IF-FOUND[, ACTION-IF-NOT-FOUND]])
 dnl
 dnl This macro looks for a library that implements the BLAS
