+++ include/Mesh.h
@@ -47,19 +47,6 @@
 #include <cmath>
 #include <stdint.h>
 
-#include <petsc-private/dmpleximpl.h>
-#include <petsc-private/isimpl.h>
-#include <petscdmda.h>
-#include <petscsf.h>
-#include <petsc.h>
-#include <petscsys.h>
-#include <petscerror.h>
-#include <petscdmplex.h>
-#include <petscviewertypes.h>
-#include <petscsf.h>
-#include <petscdm.h>
-#include "petscdmplex.h"   
-
 #ifdef HAVE_BOOST_UNORDERED_MAP_HPP
 #include <boost/unordered_map.hpp>
 #endif
@@ -85,249 +72,6 @@
 
 template<typename real_t> class Mesh{
  public:
-
-  /*! Constructor from PETSc/DMPlex
-   *
-   * @param plex DMPlex object that we import from.
-   */
-  Mesh(DM plex, MPI_Comm comm){
-    _mpi_comm = comm;
-
-    MPI_Comm_size(_mpi_comm, &num_processes);
-    MPI_Comm_rank(_mpi_comm, &rank);
-
-    // Assign the correct MPI data type to MPI_INDEX_T
-    mpi_type_wrapper<index_t> mpi_index_t_wrapper;
-    MPI_INDEX_T = mpi_index_t_wrapper.mpi_type;
-
-    nthreads = pragmatic_nthreads();
-
-    /* Establish sizes */
-    PetscErrorCode ierr;
-    PetscInt dim, cStart, cEnd, vStart, vEnd;
-    ierr = DMPlexGetDimension(plex, &dim);assert(ierr==0);
-    ierr = DMPlexGetDepthStratum(plex, 0, &vStart, &vEnd);assert(ierr==0);
-    NNodes = vEnd - vStart;
-    ierr = DMPlexGetHeightStratum(plex, 0, &cStart, &cEnd);assert(ierr==0);
-    NElements = cEnd - cStart;
-
-    if(dim == 2){
-      nloc = 3;
-      ndims = 2;
-      msize = 3;
-    }else{
-      nloc = 4;
-      ndims = 3;
-      msize = 6;
-    }
-
-    /* Pragmatic assumes that local facet numbers correspond to the local vertex number
-       of the opposite vertex. Since this is not given by the Plex we have to perform a
-       re-ordering of Plex closures in order to enforce local entity numbering according
-       to Fenics rules. The universal vertex numbering is provided by the Plex's
-       coordinate section.
-    */
-    PetscSection coord_sec;
-    ierr = DMGetCoordinateSection(plex, &coord_sec);
-    std::vector<std::pair<PetscInt, PetscInt>> fenics_local_numbering;
-    derive_fenics_local_numbering(plex, coord_sec, nloc, &fenics_local_numbering);
-
-    /* Make sure label "boundary_ids" exists */
-    PetscBool has_label;
-    ierr = DMPlexHasLabel(plex, "boundary_ids", &has_label);
-    if (!has_label) ierr = DMPlexCreateLabel(plex, "boundary_ids");
-
-    /* Import local element node list and set boundary IDs according to Plex label "boundary_ids" */
-    _ENList.resize(NElements*nloc);
-    boundary.resize(NElements*nloc);
-    int bid, ind = 0;
-    for(std::vector<std::pair<PetscInt, PetscInt>>::iterator it = fenics_local_numbering.begin();
-        it != fenics_local_numbering.end(); it++) {
-      _ENList[ind] = it->first - vStart;
-
-      ierr = DMPlexGetLabelValue(plex, "boundary_ids", it->second, &bid);
-      if (bid > 0) boundary[ind] = bid;
-      else boundary[ind] = 0;
-      ind++;
-    }
-
-    // Import local coordinates
-    _coords.resize(NNodes*ndims);
-    Vec coords_vec;
-    PetscReal *plex_coords;
-    ierr = DMGetCoordinatesLocal(plex, &coords_vec);
-    ierr = VecGetArray(coords_vec, &plex_coords);
-    /* Not sure we really need to deep copy here */
-    for (int i=0; i<NNodes*ndims; i++) _coords[i] = plex_coords[i];
-    ierr = VecRestoreArray(coords_vec, &plex_coords);
-
-    NNList.resize(NNodes);
-    NEList.resize(NNodes);
-
-    /* This std::vector< std::vector<index_t> > send, recv defines the halo
-     * where node index send[i][k] is sent to rank i, and is received by index recv[j][k], 
-     * where j is the sending rank. Note that k on the send side is the same as k on the recv size.
-     */
-    send.resize(num_processes);
-    recv.resize(num_processes);
-    send_map.resize(num_processes);
-    recv_map.resize(num_processes);
-
-    // Stores the rank that owns each vertex.
-    node_owner.resize(NNodes);
-
-    PetscSF sf = NULL;
-    PetscInt nroots, nleaves;
-    const PetscInt *local;
-    const PetscSFNode *remote;
-
-    /* Hack alert: Build a scalar CG1 section to enforce a meaningful SF */
-    PetscSection section;
-    PetscInt  comp[1] = {1};
-    PetscInt  dof[4] = {1, 0, 0, 0};
-    ierr = DMPlexCreateSection(plex, ndims, 1, comp, dof, 0, NULL, NULL, NULL, &section); assert(ierr==0);
-    ierr = DMSetDefaultSection(plex, section); assert(ierr==0);
-    ierr = DMGetDefaultSF(plex, &sf); assert(ierr==0);
-
-    /* Derive local receives, the according remote sends and a local
-       root->leaf mapping */
-    std::vector< std::vector<index_t> > remote_send(num_processes);
-    std::map<index_t, index_t> root_leaf_map;
-    ierr = PetscSFGetGraph(sf, &nroots, &nleaves, &local, &remote); assert(ierr==0);
-    assert(nleaves == NNodes);
-    for (int i=0; i<nleaves; i++) {
-      if (remote[i].rank != rank) {
-        recv[remote[i].rank].push_back( local[i] );
-        remote_send[remote[i].rank].push_back( remote[i].index );
-      } else {
-        root_leaf_map.insert( std::pair<index_t, index_t>(remote[i].index, local[i]) );
-      }
-      node_owner[i] = remote[i].rank;
-    }
-
-    /* Propagate remote send lists to the actual sender */
-    int tag = 123456;
-    std::vector<MPI_Request> send_req(num_processes);
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc != rank) {
-        ierr = MPI_Isend(remote_send[proc].data(), remote_send[proc].size(), MPI_INT,
-                         proc, tag, comm, &send_req[proc]); assert(ierr==0);
-      }
-    }
-
-    /* Receive local send list from receiving proc */
-    MPI_Status status;
-    std::vector<int> recv_size(num_processes);
-    std::vector<MPI_Request> recv_req(num_processes);
-    std::vector<index_t*> recv_buffer(num_processes);
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc != rank) {
-        ierr = MPI_Probe(proc, tag, comm, &status); assert(ierr==0);
-        ierr = MPI_Get_count(&status, MPI_INT, &recv_size[proc]); assert(ierr==0);
-        recv_buffer[proc] = new index_t[recv_size[proc]];
-
-        MPI_Irecv(recv_buffer[proc], recv_size[proc], MPI_INT, proc,
-                  tag, comm, &recv_req[proc]); assert(ierr==0);
-      }
-    }
-
-    /* Translate received send lists from roots into leaves and assign to send */
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc != rank) {
-        send[proc].resize( recv_size[proc] );
-        for (int i=0; i<recv_size[proc]; i++) {
-          send[proc][i] = root_leaf_map.find(recv_buffer[proc][i])->second;
-        }
-        free(recv_buffer[proc]);
-      }
-    }
-
-    // Stores the local2global node number map.
-    lnn2gnn.resize(NNodes);
-    ISLocalToGlobalMapping ltog;
-    const PetscInt *ltog_indices;
-    ierr = DMGetLocalToGlobalMapping(plex, &ltog); assert(ierr==0);
-    ISLocalToGlobalMappingGetIndices(ltog, &ltog_indices); assert(ierr==0);
-    for (int i=0; i<NNodes; i++) lnn2gnn[i] = ltog_indices[i];
-    ISLocalToGlobalMappingRestoreIndices(ltog, &ltog_indices); assert(ierr==0);
-
-    // Right now we are doing nothing with this. But in the future the plex may contain a special tensor field/section that we will need to import.
-    metric.resize(NNodes*msize);
-
-
-    // This really needs to be revisited...leave as it for now.
-    deferred_operations.resize(nthreads);
-#pragma omp parallel
-    {
-      // Each thread allocates nthreads DeferredOperations
-      // structs, one for each OMP thread.
-      deferred_operations[pragmatic_thread_id()].resize((defOp_scaling_factor*nthreads));
-
-#pragma omp single nowait
-      {
-        if(num_processes>1){
-          // Take into account renumbering for halo.
-          for(int j=0;j<num_processes;j++){
-            for(size_t k=0;k<recv[j].size();k++){
-              recv_halo.insert(recv[j][k]);
-            }
-            for(size_t k=0;k<send[j].size();k++){
-              send_halo.insert(send[j][k]);
-            }
-          }
-        }
-      }
-
-      // Set the orientation of elements.
-#pragma omp single
-      {
-        const int *n=get_element(0);
-        assert(n[0]>=0);
-
-        if(ndims==2)
-          property = new ElementProperty<real_t>(get_coords(n[0]),
-                                                 get_coords(n[1]),
-                                                 get_coords(n[2]));
-        else
-          property = new ElementProperty<real_t>(get_coords(n[0]),
-                                                 get_coords(n[1]),
-                                                 get_coords(n[2]),
-                                                 get_coords(n[3]));
-      }
-
-      if(ndims==2){
-#pragma omp for schedule(static)
-        for(size_t i=1;i<(size_t)NElements;i++){
-          const int *n=get_element(i);
-          assert(n[0]>=0);
-
-          double volarea = property->area(get_coords(n[0]),
-                                          get_coords(n[1]),
-                                          get_coords(n[2]));
-
-          if(volarea<0)
-            invert_element(i);
-        }
-      }else{
-#pragma omp for schedule(static)
-        for(size_t i=1;i<(size_t)NElements;i++){
-          const int *n=get_element(i);
-          assert(n[0]>=0);
-
-          double volarea = property->volume(get_coords(n[0]),
-                                            get_coords(n[1]),
-                                            get_coords(n[2]),
-                                            get_coords(n[2]));
-
-          if(volarea<0)
-            invert_element(i);
-        }
-      }
-
-      // create_adjacency is meant to be called from inside a parallel region
-      create_adjacency();
-    }
-  }
 
   /*! 2D triangular mesh constructor. This is for use when there is no MPI.
    *
@@ -460,7 +204,7 @@ template<typename real_t> class Mesh{
       }
       
       // Check neighbourhood of each element
-      for(int i=0;i<NElements;i++){
+      for(size_t i=0;i<NElements;i++){
 	if(_ENList[i*3]==-1)
 	  continue;
 	
@@ -475,7 +219,7 @@ template<typename real_t> class Mesh{
 			     inserter(neighbours, neighbours.begin()));
 	    
 	    if(neighbours.size()==2){
-	      if(*neighbours.begin()==i)
+	      if(*neighbours.begin()==(int)i)
 		boundary[i*3+j] = *neighbours.rbegin();
 	      else
 		boundary[i*3+j] = *neighbours.begin();
@@ -502,7 +246,7 @@ template<typename real_t> class Mesh{
       }
       
       // Check neighbourhood of each element
-      for(int i=0;i<NElements;i++){
+      for(size_t i=0;i<NElements;i++){
 	if(_ENList[i*4]==-1)
 	  continue;
 	
@@ -523,7 +267,7 @@ template<typename real_t> class Mesh{
 			     inserter(neighbours, neighbours.begin()));
 	    
 	    if(neighbours.size()==2){
-	      if(*neighbours.begin()==i)
+	      if(*neighbours.begin()==(int)i)
 		boundary[i*4+j] = *neighbours.rbegin();
 	      else
 		boundary[i*4+j] = *neighbours.begin();
@@ -934,40 +678,7 @@ template<typename real_t> class Mesh{
         }
       }
     }
-    return total_volume;
-  }
-
-
-  /// Get the edge length RMS value in metric space.
-  double get_lrms(){
-    double mean = get_lmean();
-
-    double rms=0;
-    int nedges=0;
-#pragma omp parallel reduction(+:rms,nedges)
-    {
-#pragma omp for schedule(static)
-      for(int i=0;i<(int)NNodes;i++){
-        if(is_owned_node(i) && (NNList[i].size()>0))
-          for(typename std::vector<index_t>::const_iterator it=NNList[i].begin();it!=NNList[i].end();++it){
-            if(i<*it){ // Ensure that every edge length is only calculated once.
-              rms+=pow(calc_edge_length(i, *it) - mean, 2);
-              nedges++;
-            }
-          }
-      }
-    }
-
-#ifdef HAVE_MPI
-    if(num_processes>1){
-      MPI_Allreduce(MPI_IN_PLACE, &rms, 1, MPI_DOUBLE, MPI_SUM, _mpi_comm);
-      MPI_Allreduce(MPI_IN_PLACE, &nedges, 1, MPI_INT, MPI_SUM, _mpi_comm);
-    }
-#endif
-
-    rms = sqrt(rms/nedges);
-
-    return rms;
+    return total_volume/6;
   }
 
   /// Get the element mean quality in metric space.
@@ -983,23 +694,24 @@ template<typename real_t> class Mesh{
         if(n[0]<0)
           continue;
 
+	double q;
         if(ndims==2){
-          sum += property->lipnikov(get_coords(n[0]), get_coords(n[1]), get_coords(n[2]),
-                                    get_metric(n[0]), get_metric(n[1]), get_metric(n[2]));
+          q = property->lipnikov(get_coords(n[0]), get_coords(n[1]), get_coords(n[2]),
+				 get_metric(n[0]), get_metric(n[1]), get_metric(n[2]));
         }else{
-          sum += property->lipnikov(get_coords(n[0]), get_coords(n[1]), get_coords(n[2]), get_coords(n[3]),
-                                    get_metric(n[0]), get_metric(n[1]), get_metric(n[2]), get_metric(n[3]));
+          q = property->lipnikov(get_coords(n[0]), get_coords(n[1]), get_coords(n[2]), get_coords(n[3]),
+				 get_metric(n[0]), get_metric(n[1]), get_metric(n[2]), get_metric(n[3]));
         }
+
+	sum+=q;
         nele++;
       }
     }
 
-#ifdef HAVE_MPI
     if(num_processes>1){
       MPI_Allreduce(MPI_IN_PLACE, &sum, 1, MPI_DOUBLE, MPI_SUM, _mpi_comm);
       MPI_Allreduce(MPI_IN_PLACE, &nele, 1, MPI_INT, MPI_SUM, _mpi_comm);
     }
-#endif
 
     double mean = sum/nele;
 
@@ -1237,52 +949,10 @@ template<typename real_t> class Mesh{
       active_vertex_map[i] = cnt++;
     }
 
-    int metis_nnodes = cnt;
-    int metis_nelements = active_element.size();
-    std::vector< std::set<index_t> > graph(metis_nnodes);
-    for(typename std::vector<index_t>::iterator ie=active_element.begin();ie!=active_element.end();++ie){
-      for(size_t i=0;i<nloc;i++){
-        index_t nid0 = active_vertex_map[_ENList[(*ie)*nloc+i]];
-        for(size_t j=i+1;j<nloc;j++){
-          index_t nid1 = active_vertex_map[_ENList[(*ie)*nloc+j]];
-          graph[nid0].insert(nid1);
-          graph[nid1].insert(nid0);
-        }
-      }
-    }
-
-    // Compress graph
-    std::vector<idxtype> xadj(metis_nnodes+1), adjncy;
-    int pos=0;
-    xadj[0]=0;
-    for(int i=0;i<metis_nnodes;i++){
-      for(typename std::set<index_t>::const_iterator jt=graph[i].begin();jt!=graph[i].end();jt++){
-        assert((*jt)>=0);
-        assert((*jt)<metis_nnodes);
-        adjncy.push_back(*jt);
-        pos++;
-      }
-      xadj[i+1] = pos;
-    }
-
-    std::vector<int> norder(metis_nnodes);
-    std::vector<int> inorder(metis_nnodes);
-
-    if(metis_nnodes != 0){
-      METIS_NodeND(&metis_nnodes, &(xadj[0]), &(adjncy[0]), NULL, NULL, &(norder[0]), &(inorder[0]));
-
-      // Update active_vertex_map
-      for(size_t i=0;i<NNodes;i++){
-        if(active_vertex_map[i]<0)
-          continue;
-
-        active_vertex_map[i] = inorder[active_vertex_map[i]];
-      }
-    }
-
     // Renumber elements
+    int active_nelements = active_element.size();
     std::map< std::set<index_t>, index_t > ordered_elements;
-    for(int i=0;i<metis_nelements;i++){
+    for(int i=0;i<active_nelements;i++){
       index_t old_eid = active_element[i];
       std::set<index_t> sorted_element;
       for(size_t j=0;j<nloc;j++){
@@ -1298,27 +968,20 @@ template<typename real_t> class Mesh{
                  <<active_vertex_map[_ENList[old_eid*nloc+2]]<<std::endl;
       }
     }
-    std::vector<index_t> metis_element_renumber;
-    metis_element_renumber.reserve(metis_nelements);
+    std::vector<index_t> element_renumber;
+    element_renumber.reserve(ordered_elements.size());
     for(typename std::map< std::set<index_t>, index_t >::const_iterator it=ordered_elements.begin();it!=ordered_elements.end();++it){
-      metis_element_renumber.push_back(it->second);
+      element_renumber.push_back(it->second);
     }
-    // assert(metis_nelements==metis_element_renumber.size());
-    metis_nelements=metis_element_renumber.size();
-    // end of renumbering
 
     // Compress data structures.
-    assert(cnt == metis_nnodes);
-    NNodes = metis_nnodes;
-    //NElements = active_element.size();
-    NElements = metis_nelements;
+    NNodes = cnt;
+    NElements = ordered_elements.size();
 
     std::vector<index_t> defrag_ENList(NElements*nloc);
     std::vector<real_t> defrag_coords(NNodes*ndims);
     std::vector<double> defrag_metric(NNodes*msize);
     std::vector<int> defrag_boundary(NElements*nloc);
-
-    assert(NElements==(size_t)metis_nelements);
 
     // This first touch is to bind memory locally.
 #pragma omp parallel
@@ -1337,8 +1000,8 @@ template<typename real_t> class Mesh{
     }
 
     // Second sweep writes elements with new numbering.
-    for(int i=0;i<metis_nelements;i++){
-      index_t old_eid = metis_element_renumber[i];
+    for(int i=0;i<NElements;i++){
+      index_t old_eid = element_renumber[i];
       index_t new_eid = i;
       for(size_t j=0;j<nloc;j++){
         index_t new_nid = active_vertex_map[_ENList[old_eid*nloc+j]];
@@ -1678,325 +1341,6 @@ template<typename real_t> class Mesh{
     }
   }
 
-  void export_dmplex(DM *plex)
-  {
-    PetscErrorCode ierr;
-    ierr = DMPlexCreateFromCellList(_mpi_comm, ndims, NElements, NNodes, nloc, PETSC_TRUE,
-                                    _ENList.data(), ndims, _coords.data(), plex); assert(ierr==0);
-
-    /* Mark boundary faces */
-    DMLabel label;
-    ierr = DMPlexCreateLabel(*plex, "boundary_faces"); assert(ierr==0);
-    ierr = DMPlexGetLabel(*plex, "boundary_faces", &label); assert(ierr==0);
-    ierr = DMPlexMarkBoundaryFaces(*plex, label); assert(ierr==0);
-
-    /* Now apply our boundary IDs to the new Plex */
-    PetscInt cell, ci, f, v, vStart, vEnd, fStart, fEnd;
-    PetscInt bid, vertex, nclosure, *closure=NULL;
-    const PetscInt *facets=NULL;
-    PetscBool incident;
-    ierr = DMPlexCreateLabel(*plex, "boundary_ids"); assert(ierr==0);
-    ierr = DMPlexGetLabel(*plex, "boundary_ids", &label); assert(ierr==0);
-    ierr = DMPlexGetDepthStratum(*plex, 0, &vStart, &vEnd);assert(ierr==0);
-    ierr = DMPlexGetHeightStratum(*plex, 1, &fStart, &fEnd);assert(ierr==0);  // facets
-    for (cell=0; cell<NElements; cell++){
-      ierr = DMPlexGetCone(*plex, cell, &facets);
-
-      for (v=0; v<nloc; v++) {
-        bid = boundary[cell*nloc+v];
-        if (bid > 0){
-          /* Find the Plex facet that is not included in the "star"
-             (inverse closure) of the non-incident vertex */
-          vertex = _ENList[cell*nloc+v] + vStart;
-          ierr = DMPlexGetTransitiveClosure(*plex, vertex, PETSC_FALSE, &nclosure, &closure);
-
-          for (f=0; f<nloc; f++) {
-            incident = PETSC_FALSE;
-            for (ci=0; ci<nclosure; ci++) {
-              if (facets[f] == closure[2*ci]) {incident=PETSC_TRUE; break;}
-            }
-            if (!incident) {
-              ierr = DMLabelSetValue(label, facets[f], bid); break;
-            }
-          }
-        }
-      }
-    }
-
-    /* After building the Plex DAG in parallel we need to build the PetscSF
-       that maps all local halo points (leaves) to remote points (roots).
-       For this we need to derive a consistent point ownership and map all
-       local plex entities to their remote roots.
-    */
-    if (num_processes <= 1) return;
-
-    PetscSection point_owners;
-    plex_create_point_ownership(*plex, &point_owners);
-
-    PetscSection remote_roots;
-    plex_map_remote_roots(*plex, &remote_roots);
-
-    PetscInt nhalo, p, pStart, pEnd, owner, root, *halo_points;
-    PetscSFNode *remote_points;
-    ierr = DMPlexGetChart(*plex, &pStart, &pEnd); assert(ierr==0);
-    nhalo = 0;
-    for (p=pStart; p<pEnd; p++) {
-      ierr = PetscSectionGetOffset(point_owners, p, &owner); assert(ierr==0);
-      if (owner != rank) nhalo++;
-    }
-
-    halo_points = new PetscInt[nhalo];
-    remote_points = new PetscSFNode[nhalo];
-    nhalo = 0;
-    for (p=pStart; p<pEnd; p++) {
-      ierr = PetscSectionGetOffset(point_owners, p, &owner); assert(ierr==0);
-      if (owner != rank) {
-        ierr = PetscSectionGetOffset(remote_roots, p, &root); assert(ierr==0);
-        halo_points[nhalo] = p;
-        remote_points[nhalo].rank = owner;
-        remote_points[nhalo].index = root;
-        nhalo++;
-      }
-    }
-    ierr = PetscSFSetGraph((*plex)->sf, pEnd-pStart, nhalo, halo_points, PETSC_OWN_POINTER,
-                           remote_points, PETSC_OWN_POINTER); assert(ierr==0);
-  }
-
-  /* This routine establishes the ownership of all points in the DAG
-     fom Pragmatic's vertex ownership. The convention is to assign a
-     point to the lowest rank of all associated vertex owners. The
-     resulting ownership is stored in a PetscSection.
-  */
-  void plex_create_point_ownership(DM plex, PetscSection *point_owners) {
-    PetscErrorCode ierr;
-    PetscInt p, pStart, pEnd, cStart, cEnd, v, vStart, vEnd;
-    PetscInt vertex, nhalo_points, owner, ci, nclosure, *closure = NULL;
-    ierr = DMPlexGetDepthStratum(plex, 0, &vStart, &vEnd);assert(ierr==0);
-    ierr = DMPlexGetHeightStratum(plex, 0, &cStart, &cEnd);assert(ierr==0);
-
-    /* Build a section to store point owners */
-    ierr = DMPlexGetChart(plex, &pStart, &pEnd); assert(ierr==0);
-    ierr = PetscSectionCreate(_mpi_comm, point_owners); assert(ierr==0);
-    ierr = PetscSectionSetChart(*point_owners, pStart, pEnd); assert(ierr==0);
-    ierr = PetscSectionSetUp(*point_owners); assert(ierr==0);
-    for (p=pStart; p<pEnd; p++) {
-      ierr = PetscSectionSetDof(*point_owners, p, 1); assert(ierr==0);
-      ierr = PetscSectionSetOffset(*point_owners, p, rank); assert(ierr==0);
-    }
-
-    /* Establish ownership of each Plex point from our halo recv lists */
-    std::vector<PetscInt> vertex_star;
-    nhalo_points = 0;
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc == rank) continue;
-
-      for (v=0; v<recv[proc].size(); v++) {
-        /* Assign ownership of halo vertex */
-        vertex = recv[proc][v] + vStart;
-        ierr = PetscSectionSetOffset(*point_owners, vertex, proc); assert(ierr==0);
-        nhalo_points++;
-
-        /* Store the vertex star (inverse closure) for later traversal */
-        ierr = DMPlexGetTransitiveClosure(plex, vertex, PETSC_FALSE, &nclosure, &closure);
-        vertex_star.resize(nclosure);
-        for (ci=0; ci<nclosure; ci++) vertex_star[ci] = closure[2*ci];
-
-        /* For each point in the vertex star determine the owner as
-           the minimum owner of all vertices in the point's closure */
-        for (p=0; p<vertex_star.size(); p++) {
-          /* Skip the original halo vertex */
-          if (vStart <= vertex_star[p] && vertex_star[p] < vEnd) continue;
-
-          owner = std::numeric_limits<int>::max();
-          ierr = DMPlexGetTransitiveClosure(plex, vertex_star[p], PETSC_TRUE, &nclosure, &closure);
-          for (ci=0; ci<nclosure; ci++) {
-            if (vStart <= closure[2*ci] && closure[2*ci] < vEnd) {
-              owner = std::min(owner, node_owner[ closure[2*ci]-vStart ]);
-            }
-          }
-          ierr = PetscSectionSetOffset(*point_owners, vertex_star[p], owner); assert(ierr==0);
-          if (owner != rank) nhalo_points++;
-        }
-      }
-    }
-  }
-
-  /* This routine matches the local plex IDs (leafs) of all points
-     in the halo/non-core region with their remote Plex IDs (roots)
-     on the sending rank.
-
-     This is done by first matching the local vertex IDs in the
-     closure(star(v)), forall v in the "send" halo via Pragmatic's GNN.
-     After that we can match entities in higher stratas (edges, facets,
-     cells) in turn by communicating the cone of the local entity and
-     matching it to the local leaf on the remote rank.
-  */
-  void plex_map_remote_roots(DM plex, PetscSection *remote_roots) {
-
-    /* Build section to store remote roots */
-    PetscErrorCode ierr;
-    PetscInt p, pStart, pEnd, cStart, cEnd, vStart, vEnd;
-    PetscInt v, ci, root, leaf, depth, size, max_cone_size, max_support_size;
-    PetscInt nmatch, njoin, *matches=NULL, nclosure, *closure=NULL;
-    const PetscInt *cone=NULL, *join=NULL;
-    std::vector< std::set<index_t> > adjacent_points(num_processes);
-    std::vector< std::set<index_t> > stratum_points(num_processes);
-    std::map<index_t,index_t> root_leaf_map;
-    std::map<index_t,index_t>::iterator entry;
-
-    /* Establish Plex sizes */
-    ierr = DMPlexGetHeightStratum(plex, 0, &cStart, &cEnd);assert(ierr==0);
-    ierr = DMPlexGetDepthStratum(plex, 0, &vStart, &vEnd);assert(ierr==0);
-    ierr = DMPlexGetMaxSizes(plex, &max_cone_size, &max_support_size); assert(ierr==0);
-    ierr = PetscMalloc(max_cone_size, &matches); assert(ierr==0);
-
-    /* Build remote_root section */
-    ierr = DMPlexGetChart(plex, &pStart, &pEnd); assert(ierr==0);
-    ierr = PetscSectionCreate(_mpi_comm, remote_roots); assert(ierr==0);
-    ierr = PetscSectionSetChart(*remote_roots, pStart, pEnd); assert(ierr==0);
-    ierr = PetscSectionSetUp(*remote_roots); assert(ierr==0);
-    for (p=pStart; p<pEnd; p++) {
-      ierr = PetscSectionSetDof(*remote_roots, p, 1); assert(ierr==0);
-      ierr = PetscSectionSetOffset(*remote_roots, p, -1); assert(ierr==0);
-    }
-
-    /* Communicate send halo vertex IDs */
-    std::vector< std::vector<index_t> > send_gnn(num_processes);
-    std::vector< std::vector<index_t> > recv_gnn(num_processes);
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc == rank) continue;
-
-      send_gnn[proc].resize( 2*send[proc].size() );
-      for (v=0; v<send[proc].size(); v++) {
-        /* Send pairs of Plex ID and GNN */
-        send_gnn[proc][2*v] = vStart + send[proc][v];
-        send_gnn[proc][2*v+1] = lnn2gnn[ send[proc][v] ];
-      }
-    }
-    send_all_to_all(send_gnn, &recv_gnn);
-
-    /* Compute the local leaf -> root mapping for "send" halo */
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc == rank) continue;
-
-      for (v=0; v<recv_gnn[proc].size(); v+=2) {
-        root = recv_gnn[proc][v];
-        leaf = vStart + recv_map[proc][ recv_gnn[proc][v+1] ];
-        ierr = PetscSectionSetOffset(*remote_roots, leaf, root); assert(ierr==0);
-        root_leaf_map.insert( std::pair<index_t,index_t>(root, leaf) );
-      }
-    }
-
-    /* Communicate recv halo vertex IDs */
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc == rank) continue;
-
-      send_gnn[proc].resize( 2*recv[proc].size() );
-      for (v=0; v<recv[proc].size(); v++) {
-        /* Send pairs of Plex ID and GNN */
-        send_gnn[proc][2*v] = vStart + recv[proc][v];
-        send_gnn[proc][2*v+1] = lnn2gnn[ recv[proc][v] ];
-      }
-    }
-    send_all_to_all(send_gnn, &recv_gnn);
-
-    /* Compute the local leaf -> root mapping for "recv" halo */
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc == rank) continue;
-
-      for (v=0; v<recv_gnn[proc].size(); v+=2) {
-        root = recv_gnn[proc][v];
-        leaf = vStart + send_map[proc][ recv_gnn[proc][v+1] ];
-        ierr = PetscSectionSetOffset(*remote_roots, leaf, root); assert(ierr==0);
-        root_leaf_map.insert( std::pair<index_t,index_t>(root, leaf) );
-      }
-    }
-
-    /* Now we can to establish the leaf->root mapping for edges, faces, etc.
-       For this we first need to derive all points adjacent to our "send"
-       vertex list by finding the closure(star(v)), where v is a "send" vertex. */
-    for (int proc=0; proc<num_processes; proc++) {
-      if (proc == rank) continue;
-
-      /* Collect all points in star(v), where v is a "send" vertex */
-      for (v=0; v<send[proc].size(); v++) {
-        ierr = DMPlexGetTransitiveClosure(plex, vStart+send[proc][v], PETSC_FALSE, &nclosure, &closure);
-        for (ci=0; ci<nclosure; ci++)  adjacent_points[proc].insert(closure[2*ci]);
-      }
-
-      /* Add the closure of all cells in star(v) */
-      for (std::set<index_t>::iterator it=adjacent_points[proc].begin();
-           it!= adjacent_points[proc].end(); it++) {
-        if (cStart <= *it && *it < cEnd) {
-          ierr = DMPlexGetTransitiveClosure(plex, *it, PETSC_TRUE, &nclosure, &closure);
-          for (ci=0; ci<nclosure; ci++)  adjacent_points[proc].insert(closure[2*ci]);
-        }
-      }
-    }
-
-    /* Finally, we can step through all layers (strata) above the vertices
-       and reconstruct the local_lnn(leaf)->remote_lnn(root) mapping. */
-    for (depth = 1; depth < ndims+1; depth++) {
-      ierr = DMPlexGetDepthStratum(plex, depth, &pStart, &pEnd);assert(ierr==0);
-      for (int proc=0; proc<num_processes; proc++) {
-        if (proc == rank) continue;
-
-        /* Go through the adjacent_points and count the entities in this stratum/layer */
-        stratum_points[proc].clear();
-        for (v=0; v<send[proc].size(); v++) {
-          for (std::set<index_t>::iterator it=adjacent_points[proc].begin();
-               it!= adjacent_points[proc].end(); it++) {
-            if (pStart <= *it && *it < pEnd) stratum_points[proc].insert(*it);
-          }
-        }
-
-        /* Pack cone of each edge touching a "send" vertex; +1 is the actual cone size */
-        send_gnn[proc].resize( stratum_points[proc].size()*(max_cone_size+2) );
-        v = 0;
-        for (std::set<index_t>::iterator it=stratum_points[proc].begin();
-             it!= stratum_points[proc].end(); it++) {
-          ierr = DMPlexGetConeSize(plex, *it, &size); assert(ierr==0);
-          ierr = DMPlexGetCone(plex, *it, &cone); assert(ierr==0);
-          send_gnn[proc][v] = *it;
-          send_gnn[proc][v+1] = size;
-          for (ci=0; ci<size; ci++)  send_gnn[proc][v+2+ci] = cone[ci];
-          v += max_cone_size + 2;
-        }
-      }
-      send_all_to_all(send_gnn, &recv_gnn);
-
-      for (int proc=0; proc<num_processes; proc++) {
-        if (proc == rank) continue;
-
-        /* Now we try to match the received cone to one of our local Plex points */
-        for (v=0; v<recv_gnn[proc].size(); v+=max_cone_size+2) {
-          root = recv_gnn[proc][v];
-          size = recv_gnn[proc][v+1];
-          nmatch = 0;
-          for (ci=2; ci<size+2; ci++) {
-            /* Go through the remote cone and find matches with locally know points */
-            entry = root_leaf_map.find( recv_gnn[proc][v+ci] );
-            if (entry != root_leaf_map.end()) {
-              matches[nmatch] = entry->second;
-              nmatch++;
-            }
-          }
-
-          /* If we know all points in the remote cone, we can find the local counterpart */
-          if (nmatch == size) {
-            ierr = DMPlexGetJoin(plex, nmatch, matches, &njoin, &join); assert(ierr==0);
-            if (njoin > 0) {
-              assert(njoin == 1);
-              ierr = PetscSectionSetOffset(*remote_roots, join[0], root); assert(ierr==0);
-              root_leaf_map.insert( std::pair<index_t,index_t>(root, join[0]) );
-            }
-          }
-        }
-      }
-    }
-    ierr = PetscFree(matches); assert(ierr==0);
-  }
-
   void send_all_to_all(std::vector< std::vector<index_t> > send_vec,
                        std::vector< std::vector<index_t> > *recv_vec) {
     int ierr, recv_size, tag = 123456;
@@ -2029,8 +1373,7 @@ template<typename real_t> class Mesh{
  private:
   template<typename _real_t> friend class MetricField2D;
   template<typename _real_t> friend class MetricField3D;
-  template<typename _real_t> friend class Smooth2D;
-  template<typename _real_t> friend class Smooth3D;
+  template<typename _real_t, int> friend class Smooth;
   template<typename _real_t> friend class Swapping2D;
   template<typename _real_t> friend class Swapping3D;
   template<typename _real_t> friend class Coarsen2D;
@@ -2668,109 +2011,6 @@ template<typename real_t> class Mesh{
         lnn2gnn[*it] = recv_buff[i][k];
     }
 #endif
-  }
-
-/* PETSc DMPlex objects do not adhere to the Fenics local numbering convention
-   for simplices when deriving the transitive closure of cells. This routine
-   performs a one time bulk re-ordering of all cell closures in a DMPlex object
-   and store the Plex points for the according vertices and facets in a
-   vector<pair<PetscInt, PetscInt>>.
-
-   Input Parameters:
-   plex -  The DM object holding mesh topology
-   vertex_numbering:
-           Section providing the global (universal) vertex numbering
-           on which the Fenics local numbering is based
-   cell_closure:
-           Transitive closures for each cell reordered accordign to Fenics
-           conventions. We allocate, but user is responsible for deallocating.
-   local_numbering:
-           Vector of size(nloc*nlements) with a pair of Plex points <vertex, facet>
-           for each nloc verteices/facets in each element according to Fenics convention.
-  */
-  void derive_fenics_local_numbering(DM plex, PetscSection vertex_numbering, size_t nloc,
-                                    std::vector<std::pair<PetscInt,PetscInt>> *local_numbering)
-  {
-    PetscInt cStart, cEnd, vStart, vEnd, fStart, fEnd;
-    PetscInt dim, cell, vertex, v_per_cell, nfacet_vertices, nclosure, *closure = NULL;
-    PetscInt f, ci, vi, fi;
-    PetscInt *vertices = NULL, *v_global = NULL, *facets = NULL, *facet_vertices = NULL;
-    PetscBool incident;
-    /* Need to do better error checking in here */
-    PetscErrorCode ierr;
-
-    ierr = DMPlexGetDimension(plex, &dim);assert(ierr==0);
-    ierr = DMPlexGetDepthStratum(plex, 0, &vStart, &vEnd);assert(ierr==0);  // vertices
-    ierr = DMPlexGetHeightStratum(plex, 1, &fStart, &fEnd);assert(ierr==0);  // facets
-    ierr = DMPlexGetHeightStratum(plex, 0, &cStart, &cEnd);assert(ierr==0);  // cells
-
-    v_per_cell = (PetscInt) nloc;
-    ierr = PetscMalloc(v_per_cell, &vertices);assert(ierr==0);
-    ierr = PetscMalloc(v_per_cell, &v_global);assert(ierr==0);
-    ierr = PetscMalloc(v_per_cell, &facets);assert(ierr==0);
-    ierr = PetscMalloc(v_per_cell - 1, &facet_vertices);assert(ierr==0);
-
-    local_numbering->resize((cEnd-cStart) * nloc);
-
-    for (cell=cStart; cell<cEnd; cell++) {
-      ierr = DMPlexGetTransitiveClosure(plex, cell, PETSC_TRUE, &nclosure, &closure);
-
-      /* Find vertices and translate universal numbers */
-      vi = 0; fi = 0;
-      for (ci=0; ci<nclosure; ci++) {
-        if (vStart <= closure[2*ci] && closure[2*ci] < vEnd) {
-          vertices[vi] = closure[2*ci];
-          ierr = PetscSectionGetOffset(vertex_numbering, closure[2*ci], &vertex);
-
-          /* Correct -ve offsets for non-owned vertices */
-          if (vertex >= 0) { v_global[vi] = vertex; }
-          else { v_global[vi] = -(vertex+1);}
-          vi++;
-        }
-
-        if (fStart <= closure[2*ci] && closure[2*ci] < fEnd) {
-          facets[fi] = closure[2*ci]; fi++;
-        }
-      }
-
-      /* Sort vertices by universal number */
-      ierr = PetscSortIntWithArray(v_per_cell, v_global, vertices);
-
-      for (vi=0; vi<v_per_cell; vi++) {
-        (*local_numbering)[cell*nloc + vi].first = vertices[vi];
-      }
-
-      for (f=0; f<nloc; f++) {
-        /* Get all vertices associated with this facet */
-        ierr = DMPlexGetTransitiveClosure(plex, facets[f], PETSC_TRUE, &nclosure, &closure);
-        vi = 0;
-        for (ci=0; ci<nclosure; ci++) {
-          if (vStart <= closure[2*ci] && closure[2*ci] < vEnd) {
-            facet_vertices[vi] = closure[2*ci]; vi++;
-          }
-        }
-        nfacet_vertices = vi;
-
-        /* Now we look for the non-incident (opposite) cell vertex */
-        for (vi=0; vi<v_per_cell; vi++) {
-          incident = PETSC_FALSE;
-          for (fi=0; fi<nfacet_vertices; fi++) {
-            if (vertices[vi] == facet_vertices[fi]) {
-              incident = PETSC_TRUE; break;
-            }
-          }
-          if (!incident) {
-            /* We have found the non-incident vertex for this facet */
-            (*local_numbering)[cell*nloc + vi].second = facets[f]; break;
-          }
-        }
-      }
-    }
-
-    PetscFree(vertices);
-    PetscFree(v_global);
-    PetscFree(facets);
-    PetscFree(facet_vertices);
   }
 
   // This is used temporarily for 3D - it will be removed in the future.
