--- petsc_makefile
@@ -1,11 +1,0 @@
-# little Makefile used to get the required LIBS and include flags 
-# out of petsc, using:
-#
-# make -f petsc_makefile getlinklibs
-# make -f petsc_makefile getincludedirs
-#
-# Before running this makefile you need to have PETSC_DIR set and
-# (unless you're using a prefix-installed installation of PETSc)
-# PETSC_ARCH
-include ${PETSC_DIR}/conf/variables
-include ${PETSC_DIR}/conf/rules
