+++ tests/src/benchmark_adapt_2d.cpp
@@ -111,7 +111,7 @@ int main(int argc, char **argv){
     double L_low = L_up/2;
 
     Coarsen2D<double> coarsen(*mesh);
-    Smooth2D<double> smooth(*mesh);
+    Smooth<double, 2> smooth(*mesh);
     Refine2D<double> refine(*mesh);
     Swapping2D<double> swapping(*mesh);
 
@@ -151,9 +151,7 @@ int main(int argc, char **argv){
       mesh->defragment();
       
       tic = get_wtime();
-      if(I>0)
-        smooth.smooth("smart Laplacian", I*10, 1.0);
-      smooth.smooth("optimisation Linf", 10);
+      smooth.smooth(10);
       toc = get_wtime();
       if(t>0) time_smooth += (toc-tic);
       if(t>0) time_adapt += (toc-tic);
